+++ docker/dev/template_dev/template.html
@@ -19,7 +19,7 @@ This Template needs excatly one Report
         <div class="relative flex flex-col min-w-0 bg-white w-full mb-6 shadow-xl rounded-lg">
           <div class="px-6">
             <div class="mt-4 mb-6">
-        <h1 class="text-4xl font-extrabold">Test Template - {{ report.date }}</h1>
+        <h1 class="text-4xl font-extrabold">SUPER Test Template - {{ report.date }}</h1>
         <h2 class="text-3xl font-semibold">{{ data.report_items[0].title }}</h2>
           Timeframe:   {{ report.timeframe }}<br/>
