+++ docker/dev/template_dev/template_dev.py
@@ -149,4 +149,4 @@ def generate(product, template) -> dict[str, bytes | str]:
 
 
 if __name__ == "__main__":
-    app.run(debug=True, port=5000)
+    app.run(debug=True, port=5000, extra_files=["template.html"])
