+++ src/core/tests/functional/config/test_config.py
